5494 Extract EnvironmentUtils
