+++ java/cdk/src/main/java/sleeper/cdk/stack/core/PropertiesStack.java
@@ -29,6 +29,7 @@ import sleeper.cdk.stack.core.LoggingStack.LogGroupRef;
 import sleeper.cdk.util.Utils;
 import sleeper.core.deploy.LambdaHandler;
 import sleeper.core.properties.instance.InstanceProperties;
+import sleeper.core.util.EnvironmentUtils;
 
 import java.util.Map;
 
@@ -53,7 +54,7 @@ public class PropertiesStack extends NestedStack {
         IFunction propertiesWriterLambda = lambdaCode.buildFunction(this, LambdaHandler.PROPERTIES_WRITER, "PropertiesWriterLambda", builder -> builder
                 .functionName(functionName)
                 .memorySize(2048)
-                .environment(Utils.createDefaultEnvironment(instanceProperties))
+                .environment(EnvironmentUtils.createDefaultEnvironment(instanceProperties))
                 .description("Lambda for writing instance properties to S3 upon initialisation and teardown")
                 .logGroup(coreStacks.getLogGroup(LogGroupRef.PROPERTIES_WRITER)));
